Added Core Data Entities
Added entities for tweet and user
added son parsing logic to create core data entities.
