+++ TwitterSearch/Base.lproj/Main.storyboard
@@ -9,7 +9,7 @@
         <!--Twitter Search View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController useStoryboardIdentifierAsRestorationIdentifier="YES" id="BYZ-38-t0r" customClass="TwitterSearchViewController" customModule="TwitterSearch" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController storyboardIdentifier="TwitterSearchViewController" useStoryboardIdentifierAsRestorationIdentifier="YES" id="BYZ-38-t0r" customClass="TwitterSearchViewController" customModule="TwitterSearch" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
